Hands/Arms Free Mask Applicator
This device is a mask applicator that requires no use of the hands or arms. It is designed to be mounted to a wall. By bringing one's head into the device, using their neck strength, they will draw the two lever arms around their head, thereby placing the mask on their face.

Contained in the file are preliminary sketches and a homemade prototype.